Set query on clicking list item
