+++ src/app/views/sidebar/resource-explorer/ResourceExplorer.tsx
@@ -7,10 +7,13 @@ import React, { useEffect, useState } from 'react';
 import { FormattedMessage } from 'react-intl';
 import { useDispatch, useSelector } from 'react-redux';
 import { telemetry, eventTypes, componentNames } from '../../../../telemetry';
+import { IQuery } from '../../../../types/query-runner';
 
-import { IResource, IResourceLink, ResourceOptions } from '../../../../types/resources';
+import { IResource, IResourceLink, ResourceLinkType, ResourceOptions } from '../../../../types/resources';
 import { IRootState } from '../../../../types/root';
+import { setSampleQuery } from '../../../services/actions/query-input-action-creators';
 import { addResourcePaths } from '../../../services/actions/resource-explorer-action-creators';
+import { GRAPH_URL } from '../../../services/graph-constants';
 import { translateMessage } from '../../../utils/translate-messages';
 import { classNames } from '../../classnames';
 import { sidebarStyles } from '../Sidebar.styles';
@@ -143,6 +146,7 @@ const unstyledResourceExplorer = (props: any) => {
   const clickLink = (ev?: React.MouseEvent<HTMLElement>, item? : INavLink) => {
     ev!.preventDefault();
     item!.isExpanded = !item!.isExpanded;
+    setQuery(item!);
   }
 
   const resourceOptionSelected = (activity: string, context: any) => {
@@ -150,6 +154,27 @@ const unstyledResourceExplorer = (props: any) => {
       addToCollection(context);
     }
   }
+
+  const setQuery = (resourceLink: INavLink) => {
+    if (resourceLink.type === ResourceLinkType.NODE) { return; }
+    const resourceUrl = getUrlFromLink(resourceLink);
+    if (!resourceUrl) { return; }
+    const sampleUrl = `${GRAPH_URL}/${version}${resourceUrl}`;
+    const query: IQuery = {
+      selectedVerb: resourceLink.method!,
+      selectedVersion: version,
+      sampleUrl,
+      sampleHeaders: [],
+      sampleBody: undefined
+    };
+    dispatch(setSampleQuery(query));
+    telemetry.trackEvent(eventTypes.LISTITEM_CLICK_EVENT, {
+      ComponentName: componentNames.RESOURCES_LIST_ITEM,
+      ResourceLink: resourceUrl,
+      SelectedVersion: version
+    });
+  }
+
 
   const breadCrumbs = generateBreadCrumbs();
